+++ src/trader/application/training/__init__.py
@@ -1,0 +1,3 @@
+"""Application boundaries for profile-owned offline training."""
+
+__all__: tuple[str, ...] = ()